Handle artifact/capp deployment behavior when hot deployement is disabled
