+++ components/org.wso2.micro.integrator.initializer/pom.xml
@@ -92,6 +92,10 @@
             <groupId>org.wso2.carbon.mediation</groupId>
             <artifactId>org.wso2.carbon.mediation.library</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.ei</groupId>
+            <artifactId>org.wso2.micro.integrator.event.sink</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
