+++ components/org.wso2.micro.integrator.initializer/src/main/java/org/wso2/micro/integrator/initializer/deployment/AppDeployerServiceComponent.java
@@ -17,17 +17,12 @@
  */
 package org.wso2.micro.integrator.initializer.deployment;
 
-import java.io.File;
-
 import org.apache.axis2.context.ConfigurationContext;
-import org.apache.axis2.deployment.Deployer;
 import org.apache.axis2.deployment.DeploymentEngine;
-import org.apache.axis2.deployment.DeploymentException;
-import org.apache.axis2.description.Parameter;
 import org.apache.axis2.engine.AxisConfigurator;
-import org.apache.axis2.util.JavaUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.wso2.micro.integrator.event.sink.EventSinkDeployer;
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
@@ -37,17 +32,15 @@ import org.osgi.service.component.annotations.ReferenceCardinality;
 import org.osgi.service.component.annotations.ReferencePolicy;
 import org.wso2.micro.application.deployer.handler.DefaultAppDeployer;
 import org.wso2.micro.core.CarbonAxisConfigurator;
-import org.wso2.micro.core.util.CarbonException;
 import org.wso2.micro.integrator.dataservices.core.DBDeployer;
+import org.wso2.micro.integrator.initializer.ServiceBusConstants;
 import org.wso2.micro.integrator.initializer.StartupFinalizer;
-import org.wso2.micro.integrator.initializer.deployment.application.deployer.CAppDeploymentManager;
-import org.wso2.micro.integrator.initializer.deployment.artifact.deployer.ArtifactDeploymentManager;
+import org.wso2.micro.integrator.initializer.deployment.application.deployer.CappDeployer;
 import org.wso2.micro.integrator.initializer.deployment.synapse.deployer.FileRegistryResourceDeployer;
 import org.wso2.micro.integrator.initializer.deployment.synapse.deployer.SynapseAppDeployer;
 import org.wso2.micro.integrator.initializer.services.SynapseEnvironmentService;
 import org.wso2.micro.integrator.initializer.utils.ConfigurationHolder;
 import org.wso2.micro.integrator.ndatasource.capp.deployer.DataSourceCappDeployer;
-import static org.wso2.micro.integrator.initializer.deployment.DeploymentConstants.TAG_HOT_DEPLOYMENT;
 
 @Component(name = "org.wso2.micro.integrator.initializer.deployment.AppDeployerServiceComponent", immediate = true)
 public class AppDeployerServiceComponent {
@@ -62,48 +55,28 @@ public class AppDeployerServiceComponent {
     protected void activate(ComponentContext ctxt) {
 
         if (log.isDebugEnabled()) {
-            log.debug(AppDeployerServiceComponent.class.getName() + "#activate() BEGIN - " + System.currentTimeMillis());
+            log.debug(
+                    AppDeployerServiceComponent.class.getName() + "#activate() BEGIN - " + System.currentTimeMillis());
             log.debug("Activating AppDeployerServiceComponent");
         }
 
         // ConfigurationHolder is updated by org.wso2.micro.integrator.initializer.ServiceBusInitializer
         configCtx = ConfigurationHolder.getInstance().getAxis2ConfigurationContextService().getServerConfigContext();
 
-        // Initialize deployers
-        ArtifactDeploymentManager artifactDeploymentManager = new ArtifactDeploymentManager(configCtx.getAxisConfiguration());
-        CAppDeploymentManager cAppDeploymentManager = CAppDeploymentManager.getInstance();
-        cAppDeploymentManager.init(configCtx.getAxisConfiguration());
+        // Initialize deployers in micro integrator
+        initializeDeployers();
 
-        initializeDeployers(artifactDeploymentManager, cAppDeploymentManager);
-
-        // Register eventSink deployer
-        registerEventSinkDeployer(artifactDeploymentManager);
-
-        // Deploy artifacts
-        artifactDeploymentManager.deploy();
-
-        Parameter hotDeployment = configCtx.getAxisConfiguration().getParameter(TAG_HOT_DEPLOYMENT);
-        if (hotDeployment != null && JavaUtils.isTrue(hotDeployment.getValue(), true)) {
-            // Invoke registered deployers, so that they will be triggered followed by hot deployment.
-            invokeRegisteredDeployers();
-        } else {
-            try {
-                // Deploy all carbon applications in cApp directory during the startup time.
-                cAppDeploymentManager.deploy();
-            } catch (CarbonException e) {
-                log.error("Error occurred while deploying carbon application", e);
-            }
-            log.debug("MicroIntegrator artifact/Capp Deployment completed");
-        }
+        // Invoke all registered deployers to deploy services
+        invokeRegisteredDeployers();
 
         // Finalize server startup
         startupFinalizer = new StartupFinalizer(configCtx, ctxt.getBundleContext());
         startupFinalizer.finalizeStartup();
 
         if (log.isDebugEnabled()) {
-            log.debug(AppDeployerServiceComponent.class.getName() + "#activate() COMPLETE - " + System.currentTimeMillis());
+            log.debug(AppDeployerServiceComponent.class.getName() + "#activate() COMPLETE - " +
+                              System.currentTimeMillis());
         }
-
     }
 
     @Deactivate
@@ -156,95 +129,82 @@ public class AppDeployerServiceComponent {
     }
 
     /**
-     * Function to initialize deployer
-     *
-     * @param artifactDeploymentManager
-     * @param cAppDeploymentManager
+     * Function to initialize deployers.
      */
-    private void initializeDeployers(ArtifactDeploymentManager artifactDeploymentManager,
-                                     CAppDeploymentManager cAppDeploymentManager) {
-
-        String artifactRepoPath = configCtx.getAxisConfiguration().getRepository().getPath();
-
-        log.debug("Initializing ArtifactDeploymentManager deployment manager");
-
-        // TODO :- unComment after installing DSS
-        // Create data services deployer
-        DBDeployer dbDeployer = new DBDeployer();
-        dbDeployer.setDirectory(artifactRepoPath + DeploymentConstants.DSS_DIR_NAME);
-        dbDeployer.setExtension(DeploymentConstants.DSS_TYPE_EXTENSION);
-
-        // Register artifact deployers in ArtifactDeploymentManager
-        try {
-            artifactDeploymentManager.registerDeployer(artifactRepoPath + DeploymentConstants.DSS_DIR_NAME, dbDeployer);
-        } catch (DeploymentException e) {
-            log.error("Error occurred while registering data services deployer");
-        }
-
-        // Initialize micro integrator carbon application deployer
-        log.debug("Initializing carbon application deployment manager");
-
-        // Register deployers in DeploymentEngine (required for CApp deployment)
+    private void initializeDeployers() {
         DeploymentEngine deploymentEngine = (DeploymentEngine) configCtx.getAxisConfiguration().getConfigurator();
-        deploymentEngine.addDeployer(dbDeployer, DeploymentConstants.DSS_DIR_NAME, DeploymentConstants.DSS_TYPE_DBS);
 
-        // Register application deployment handlers
-        //TODO
-       cAppDeploymentManager.registerDeploymentHandler(new FileRegistryResourceDeployer(
-                synapseEnvironmentService.getSynapseEnvironment().getSynapseConfiguration().getRegistry()));
-        cAppDeploymentManager.registerDeploymentHandler(new DataSourceCappDeployer());
-        cAppDeploymentManager.registerDeploymentHandler(new DefaultAppDeployer());
-        cAppDeploymentManager.registerDeploymentHandler(new SynapseAppDeployer());
+        // Register data services deployer in DeploymentEngine
+        addDBDeployer(deploymentEngine);
 
         // Register CappDeployer in DeploymentEngine (required for CApp hot deployment)
-        addCAppDeployer();
+        addCAppDeployer(deploymentEngine);
+
+        // Register EventSinkDeployer in DeploymentEngine
+        addEventSinkDeployer(deploymentEngine);
     }
 
     /**
-     * Function to register eventSink deployer.
-     *
-     * @param artifactDeploymentManager artifactDeploymentManager which manages and performs artifact deployment
+     * Initialize and add the EventSinkDeployer to the Deployment Engine.
      */
-    public void registerEventSinkDeployer(ArtifactDeploymentManager artifactDeploymentManager) {
-        try {
-            String carbonRepoPath = configCtx.getAxisConfiguration().getRepository().getPath();
-            String eventSinkPath = carbonRepoPath + File.separator + "event-sinks";
-            Class deployerClass = Class.forName("org.wso2.micro.integrator.event.sink.EventSinkDeployer");
-            Deployer deployer = (Deployer) deployerClass.newInstance();
-            artifactDeploymentManager.registerDeployer(eventSinkPath, deployer);
-            if (log.isDebugEnabled()) {
-                log.debug("Successfully registered eventSink deployer");
-            }
-        } catch (ClassNotFoundException e) {
-            log.error("Can not find class EventSinkDeployer", e);
-        } catch (InstantiationException e) {
-            log.error("Error instantiating EventSinkDeployer class", e);
-        } catch (DeploymentException e) {
-            log.error("Error registering eventSink deployer", e);
-        } catch (IllegalAccessException e) {
-            log.error("Error instantiating EventSinkDeployer class", e);
+    private void addEventSinkDeployer(DeploymentEngine deploymentEngine) {
+        String artifactRepoPath = configCtx.getAxisConfiguration().getRepository().getPath();
+
+        // Initialize EventSource Deployer here
+        EventSinkDeployer eventSinkDeployer = new EventSinkDeployer();
+        eventSinkDeployer.setDirectory(artifactRepoPath + DeploymentConstants.EVENT_SINK_DIR_NAME);
+
+        //Add the deployer to the deployment engine
+        deploymentEngine.addDeployer(eventSinkDeployer, DeploymentConstants.EVENT_SINK_DIR_NAME,
+                                     ServiceBusConstants.ARTIFACT_EXTENSION);
+
+        if (log.isDebugEnabled()) {
+            log.debug("Successfully registered eventSink deployer");
         }
     }
 
     /**
      * Initialize and add the CappDeployer to the Deployment Engine.
      */
-    private void addCAppDeployer() {
-        try {
-            String appsRepo = "carbonapps";
-            // Initialize CApp deployer here
-            Class deployerClass = Class.
-                    forName("org.wso2.micro.integrator.initializer.deployment.application.deployer.CappDeployer");
-            Deployer deployer = (Deployer) deployerClass.newInstance();
+    private void addCAppDeployer(DeploymentEngine deploymentEngine) {
+        String artifactRepoPath = configCtx.getAxisConfiguration().getRepository().getPath();
 
-            //Add the deployer to deployment engine
-            DeploymentEngine deploymentEngine =
-                    (DeploymentEngine) configCtx.getAxisConfiguration().getConfigurator();
-            deploymentEngine.addDeployer(deployer, appsRepo, "car");
-        } catch (ClassNotFoundException e) {
-            log.error("Can not find class CappDeployer", e);
-        } catch (IllegalAccessException | InstantiationException e) {
-            log.error("Error while instantiating CappDeployer class", e);
+        // Initialize CApp deployer here
+        CappDeployer cappDeployer = new CappDeployer();
+        cappDeployer.setDirectory(artifactRepoPath + DeploymentConstants.CAPP_DIR_NAME);
+        cappDeployer.init(configCtx);
+
+        // Register application deployment handlers
+        cappDeployer.registerDeploymentHandler(new FileRegistryResourceDeployer(
+                synapseEnvironmentService.getSynapseEnvironment().getSynapseConfiguration().getRegistry()));
+        cappDeployer.registerDeploymentHandler(new DataSourceCappDeployer());
+        cappDeployer.registerDeploymentHandler(new DefaultAppDeployer());
+        cappDeployer.registerDeploymentHandler(new SynapseAppDeployer(configCtx.getAxisConfiguration()));
+
+        //Add the deployer to deployment engine. This should be done after registering the deployment handlers.
+        deploymentEngine.addDeployer(cappDeployer, artifactRepoPath + DeploymentConstants.CAPP_DIR_NAME,
+                                     DeploymentConstants.CAPP_TYPE_EXTENSION);
+        if (log.isDebugEnabled()) {
+            log.debug("Successfully registered CappDeployer");
+        }
+    }
+
+    /**
+     * Initialize and add the Data Service Deployer to the Deployment Engine.
+     */
+    private void addDBDeployer(DeploymentEngine deploymentEngine) {
+        String artifactRepoPath = configCtx.getAxisConfiguration().getRepository().getPath();
+
+        // Create data services deployer
+        DBDeployer dbDeployer = new DBDeployer();
+        dbDeployer.setDirectory(artifactRepoPath + DeploymentConstants.DSS_DIR_NAME);
+        dbDeployer.setExtension(DeploymentConstants.DSS_TYPE_EXTENSION);
+
+        // Register deployer in DeploymentEngine
+        deploymentEngine.addDeployer(dbDeployer, DeploymentConstants.DSS_DIR_NAME, DeploymentConstants.DSS_TYPE_DBS);
+
+        if (log.isDebugEnabled()) {
+            log.debug("Successfully registered Data Service Deployer");
         }
     }
 
@@ -252,13 +212,9 @@ public class AppDeployerServiceComponent {
      * Invoke all registered deployers.
      */
     private void invokeRegisteredDeployers() {
-        try {
-            AxisConfigurator axisConfigurator = configCtx.getAxisConfiguration().getConfigurator();
-            if (axisConfigurator instanceof CarbonAxisConfigurator) {
-                ((CarbonAxisConfigurator) axisConfigurator).deployServices();
-            }
-        } catch (Exception e) {
-            log.error("Runtime exception while deploying artifacts ", e);
+        AxisConfigurator axisConfigurator = configCtx.getAxisConfiguration().getConfigurator();
+        if (axisConfigurator instanceof CarbonAxisConfigurator) {
+            ((CarbonAxisConfigurator) axisConfigurator).deployServices();
         }
     }
 }
